--- a/Cargo.toml
+++ b/Cargo.toml
@@ -6,14 +6,15 @@ edition = "2024"
 [dependencies]
 fst = "0.4"
 okkhor = { version = "0.7", features = ["regex"] }
-regex = "1"
-regex-automata = { version = "0.4" }
+regex-automata = { version = "0.4", features = ["dfa-search"]}
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 
 [dev-dependencies]
-criterion = { version = "0.5", features = ["html_reports"] }
+criterion = { version = "0.6", features = ["html_reports"] }
 ahash = "0.8"
+peak_alloc = "0.2"
+regex = "1"
 
 [[bench]]
 name = "suggestions"