--- a/src/lib.rs
+++ b/src/lib.rs
@@ -1,3 +1,4 @@
 mod trie;
 mod utils;
 pub mod suggest;
+pub mod regex;