--- a/src/lib.rs
+++ b/src/lib.rs
@@ -1,4 +1,5 @@
 mod trie;
 mod utils;
+mod automation;
 pub mod suggest;
 pub mod regex;