This was an attempt to use the `fst` crate to represent the dictionary and search the fst using regex. The `fst` provides significant memory usage reduction; only 1 MB is needed to represent the 3.8 MB dictionary compared to the current Trie representation, which uses 145 MB. But the downside is that searching with regexes is tremendously slow compared to current implementation with trie based search.

Mode | upodesh Time (µs) | upodesh-fst Time (µs) | Slowdown Factor (fst vs baseline) 
-- | -- | -- | -- 
a | 2.1615 | 104.86 | ~48.5× slower 
arO | 7.0134 | 261.75 | ~37.3× slower 
bistari | 4.6440 | 402.60 | ~86.6× slower 

<details>
upodesh a               time:   [2.1604 µs 2.1615 µs 2.1625 µs]
                        change: [−20.775% −20.677% −20.575%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  3 (3.00%) low severe
  3 (3.00%) high mild
  4 (4.00%) high severe

upodesh arO             time:   [7.0103 µs 7.0134 µs 7.0167 µs]
                        change: [−27.864% −27.793% −27.726%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  2 (2.00%) high mild
  5 (5.00%) high severe

upodesh bistari         time:   [4.6418 µs 4.6440 µs 4.6464 µs]
                        change: [−24.130% −24.050% −23.972%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe

upodesh-fst a           time:   [104.81 µs 104.86 µs 104.92 µs]
Found 9 outliers among 100 measurements (9.00%)
  1 (1.00%) low mild
  4 (4.00%) high mild
  4 (4.00%) high severe

upodesh-fst arO         time:   [261.58 µs 261.75 µs 261.93 µs]
Found 11 outliers among 100 measurements (11.00%)
  6 (6.00%) high mild
  5 (5.00%) high severe

upodesh-fst bistari     time:   [400.92 µs 402.60 µs 404.08 µs]

regex a                 time:   [195.83 µs 195.90 µs 195.98 µs]
                        change: [+0.9174% +1.0273% +1.1363%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe

regex arO               time:   [247.07 µs 247.50 µs 247.93 µs]
                        change: [−0.4496% −0.2532% −0.0679%] (p = 0.01 < 0.05)
                        Change within noise threshold.
Found 9 outliers among 100 measurements (9.00%)
  8 (8.00%) high mild
  1 (1.00%) high severe

regex bistari           time:   [358.10 µs 358.50 µs 358.85 µs]
                        change: [+0.7364% +0.9580% +1.1805%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe
</details>


#3